// src/app/models/search-filters.ts
export class SearchFilters {

  city: string;
  startDate: string;
  endDate: string;
  minAmount = 0;
  maxAmount = 1000;
  ascending = true;
  itemOnPage = 10;
  currentPage = 1;
  make: string;
  model: string;
  year: string;
  engine: string;
  hoursepower: string;
  fuelconsumption: string;
  wd: string;
  transmission: string;
  fuel: string;
}